update desafio CRUD

+++ 1-java-spring-essential/desafio-CRUD-clientes/src/main/java/com/devsuperior/desafioCRUDclientes/services/ClientService.java
@@ -4,14 +4,13 @@ import com.devsuperior.desafioCRUDclientes.dto.ClientDTO;
 import com.devsuperior.desafioCRUDclientes.entities.Client;
 import com.devsuperior.desafioCRUDclientes.repositories.ClientRepository;
 import com.devsuperior.desafioCRUDclientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ClientService {
@@ -40,15 +39,26 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO clientDTO) {
-        Client client = clientRepository.getReferenceById(id);
-        copyDtoToEntity(clientDTO, client);
-        client = clientRepository.save(client);
-        return new ClientDTO(client);
+        try {
+            Client client = clientRepository.getReferenceById(id);
+            copyDtoToEntity(clientDTO, client);
+            client = clientRepository.save(client);
+            return new ClientDTO(client);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("ID não encontrado");
+        }
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("ID não encontrado");
+        }
+        else {
+            clientRepository.deleteById(id);
+        }
+
     }
+
     private void copyDtoToEntity(ClientDTO clientDTO, Client client) {
         client.setName(clientDTO.getName());
         client.setCpf(clientDTO.getCpf());
